+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * uses quickjs_runtime 0.10.0 and ditches the multi engine principles
 * removed console in favour of impl in quickjs_runtime
 * added opt features for console/settimeout/setinterval/setimmediate
+* added support for dataset/style objs on Elements in htmldom
 
 # 0.0.4